# backend/docs/swagger.yaml
basePath: /
definitions:
...
  groups.addMemberRequest:
    properties:
      display_name:
        type: string
      is_leader:
        type: boolean
      plaid_access_token:
        type: string
      plaid_account_id:
        type: string
      split_weight:
        type: number
    required:
    - display_name
    type: object
  groups.addMemberResponse:
    properties:
      display_name:
        type: string
      member_id:
        type: string
      split_weight:
        type: number
      user_id:
        type: string
    type: object
  groups.createGroupRequest:
    properties:
      currency:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  groups.createGroupResponse:
    properties:
      created_at:
        type: string
      currency:
        type: string
      group_id:
        type: string
      name:
        type: string
    type: object
  groups.getGroupResponse:
    properties:
      currency:
        type: string
      group_id:
        type: string
      members:
        items:
          $ref: '#/definitions/groups.memberSummary'
        type: array
      name:
        type: string
    type: object
  groups.memberSummary:
    properties:
      display_name:
        type: string
      has_card:
        type: boolean
      is_leader:
        type: boolean
      member_id:
        type: string
      split_weight:
        type: number
      tally_balance_cents:
        type: integer
    type: object
  groups.transactionSummary:
    properties:
      amount_cents:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      merchant_category:
        type: string
      merchant_name:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend for Tally — a group card-spending app with JIT authorization
    and a double-entry ledger.
  title: Tally API
  version: "1.0"
paths:
...
  /v1/groups:
    post:
      consumes:
      - application/json
      description: Creates a new Tally group and provisions its double-entry liability
        ledger account.
      parameters:
      - description: Group details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/groups.createGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/groups.createGroupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a group
      tags:
      - groups
  /v1/groups/{id}:
    get:
      description: Returns group metadata (name, currency) and a summary of all members
        including balances.
      parameters:
      - description: Group ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.getGroupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a group
      tags:
      - groups
  /v1/groups/{id}/members:
    post:
      consumes:
      - application/json
      description: Creates a member record (with optional Plaid tokens) and provisions
        their asset ledger account.
      parameters:
      - description: Group ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/groups.addMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/groups.addMemberResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a member to a group
      tags:
      - groups
  /v1/groups/{id}/transactions:
    get:
      description: Returns the 50 most recent transactions for a group, newest first.
      parameters:
      - description: Group ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/groups.transactionSummary'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List transactions
      tags:
      - groups
...
swagger: "2.0"